+++ local-examples/components/app.js
@@ -88,186 +88,42 @@
   app.controller('MainCtrl', function MainCtrl($timeout, $q) {
     var vm = this;
 
-    vm.user = {};
+    vm.formData = {
+          "name": "John",
+          "address": {
+              "streetnum": "145 Elm St.",
+              "city": "Atlanta"
+          }
+    };
 
-    vm.swapFields = swapFields;
-    let key = 0;
-
-    vm.realFields = [
+    vm.formFields = [
       {
-        elementAttributes: {
-          layout: ''
-        },
-        className: 'display-flex',
-        fieldGroup: [
-          {elementAttributes: {flex: ''}, className: 'flex-1', type: 'input', key: key++, templateOptions: {label: 'Field ' + key}},
-          {className: 'flex-2', type: 'input', key: key++, templateOptions: {label: 'Field ' + key}}
-        ]
+        key: 'name',
+        type: 'input',
+        templateOptions: {
+          label: 'Name'
+        }
       },
-      {type: 'input', key: key++, templateOptions: {label: 'Field ' + key}},
-      {type: 'input', key: key++, templateOptions: {label: 'Field ' + key}},
       {
+        key: 'address',
         fieldGroup: [
-          {type: 'input', key: key++, templateOptions: {label: 'Field ' + key}},
-          {type: 'input', key: key++, templateOptions: {label: 'Field ' + key}},
-          {type: 'input', key: key++, templateOptions: {label: 'Field ' + key}}
+          {
+            key: 'streetnum',
+            type: 'input',
+            templateOptions: {
+              label: 'Street Num'
+            }
+          },
+          {
+            key: 'city',
+            type: 'input',
+            templateOptions: {
+              label: 'City'
+            }
+          }
         ]
       }
     ];
-
-    vm.fields = getFields();
-
-    function swapFields() {
-      vm.fields = getFields();
-    }
-
-    function getFields() {
-      return [
-        {
-          type: 'checkbox',
-          key: 'mine',
-          templateOptions: {
-            label: 'My Label',
-            description: 'This is an awesome description for a checkbox!',
-            required: true
-          }
-        },
-        {
-          type: 'input',
-          key: 'maxLength',
-          templateOptions: {
-            type: 'number',
-            label: 'Max Length',
-            onChange: function(value, options, scope) {
-              scope.formState.maxLength = value;
-            }
-          }
-        },
-        {
-          type: 'input',
-          key: 'myKey',
-          templateOptions: {
-            placeholder: 'This rocks',
-            label: 'My Input',
-            required: true,
-            description: 'This is an awesome description',
-            focus: true,
-            addonRight: {
-              class: 'glyphicon glyphicon-ok'
-            }
-          },
-          expressionProperties: {
-            'templateOptions.label': '$viewValue',
-            'templateOptions.maxlength': 'formState.maxLength',
-            'templateOptions.tabindex': 'model.mine ? 0 : -1'
-          }
-        },
-        {
-          type: 'select',
-          key: 'mySelect',
-          templateOptions: {
-            label: 'Choose something!',
-            options: [
-              {},
-              {display: 'item 1', id: 'coolio'},
-              {display: 'item 2', id: 'coolio2'},
-              {display: 'item 3', id: 'coolio3'}
-            ],
-            valueProp: 'id',
-            labelProp: 'display'
-          }
-        },
-        {
-          type: 'multiCheckbox',
-          key: 'multipleOptions',
-          templateOptions: {
-            label: 'Multiple Options',
-            options: [
-              {label: 'Cool cat', value: {a: 'b'}},
-              {label: 'Cool Dog', value: {c: 'd'}},
-              {label: 'Cool Frog', value: true},
-              {label: 'Cool Log', value: 34}
-            ],
-            labelProp: 'label'
-          }
-        },
-        {
-          type: 'customExtended',
-          key: 'myCustomThing',
-          templateOptions: {
-            label: 'Custom stuff',
-            description: 'This has a link and controller!'
-          },
-          link: function(scope, el) {
-            console.log(scope, el);
-          },
-          controller: function($scope, $log) {
-            $log.info($scope);
-          }
-        },
-        {
-          type: 'textarea',
-          key: 'coolTextarea',
-          templateOptions: {
-            label: 'Type stuff',
-            placeholder: 'Way fun',
-            cols: 15,
-            rows: 12
-          }
-        },
-        {
-          type: 'textarea',
-          key: 'coolTextarea',
-          templateOptions: {
-            label: 'Type stuff',
-            placeholder: 'Way fun'
-          }
-        },
-        {
-          type: 'radio',
-          key: 'myRadios',
-          templateOptions: {
-            label: 'Cool Radios',
-            disabled: true,
-            options: [
-              {name: 'item 1', value: 'coolio'},
-              {name: 'item 2', value: 'coolio2'},
-              {name: 'item 3', value: 'coolio3'}
-            ],
-            description: 'Click one!'
-          }
-        },
-        {
-          type: 'input',
-          key: 'aNumber',
-          templateOptions: {
-            type: 'number',
-            label: 'Number stuff',
-            placeholder: '10 is the max, -10 is the min...',
-            description: null,
-            max: 10,
-            min: -10
-          }
-        },
-        {
-          type: 'input',
-          key: 'email',
-          validators: {
-            specialEmail: function(modelValue, viewValue) {
-              return $timeout(function() {
-                var fn = (modelValue || viewValue) === 'a@b.c' ? $q.when : $q.reject;
-                return fn();
-              }, 2000);
-            }
-          },
-          templateOptions: {
-            type: 'email',
-            placeholder: 'Type a@b.c if you want to be valid',
-            label: 'Special Email address'
-          }
-        }
-      ];
-    }
   });
 
 })();