+++ src/providers/formlyApiCheck.js
@@ -63,7 +63,7 @@ const expressionProperties = apiCheck.objectOf(apiCheck.oneOfType([
   }).strict
 ]));
 
-const modelChecker = apiCheck.oneOfType([apiCheck.oneOf(['formState']), apiCheck.object]);
+const modelChecker = apiCheck.oneOfType([apiCheck.oneOf(['formState']), apiCheck.object, apiCheck.string]);
 
 let fieldOptionsApiShape = {
   $$hashKey: apiCheck.any.optional,
@@ -156,6 +156,7 @@ const fieldGroup = apiCheck.shape({
   options: formOptionsApi.optional,
   hide: apiCheck.bool.optional,
   hideExpression: formlyExpression.optional,
+  key: apiCheck.oneOfType([apiCheck.string, apiCheck.number]).optional,
   model: modelChecker.optional,
   form: apiCheck.object.optional,
   elementAttributes: apiCheck.objectOf(apiCheck.string).optional